fixup: Moved Comm::Descriptor to fd.h for future reuse
